+++ frontend/src/scenes/experiments/charts/funnels/StepBarLabels.tsx
