+++ frontend/src/scenes/experiments/charts/funnels/StepLegend.tsx
